paralegal_rustc_utils::mir::place
Struct RegionVisitor

struct RegionVisitor<'tcx, Dispatcher> {
    tcx: TyCtxt<'tcx>,
    def_id: DefId,
    local: Local,
    place_stack: Vec<PlaceElem<'tcx>>,
    ty_stack: Vec<Ty<'tcx>>,
    dispatcher: Dispatcher,
    stop_at: StoppingCondition,
}

Fields
tcx: TyCtxt<'tcx>def_id: DefIdlocal: Local

Base local of the place we are collecting regions for.

place_stack: Vec<PlaceElem<'tcx>>

List of projections to apply to the base local in order to reach the child place currently under consideration.

Starts out as the projections in the input place.

ty_stack: Vec<Ty<'tcx>>

Sequence of parent types to reach the place currently under consideration. Correspond to the projections in place_stack.

dispatcher: Dispatcher

Callbacks

stop_at: StoppingCondition
Implementations

impl<'tcx, Dispatcher: Default> RegionVisitor<'tcx, Dispatcher>

fn new( tcx: TyCtxt<'tcx>, def_id: DefId, place: Place<'tcx>, stop_at: StoppingCondition, ) -> Self

Construct a new [CollectRegions] visitor.

By itself the visitor only implements the traversal. Actual accumulation of useful information is done by the Dispatcher.

fn into_inner(self) -> Dispatcher